Handle a grey area - few sentences but many tokens
The minimum number of sentences to cluster is 22, since we reduce dimensionality to 20 using UMAP

If we have <22 sentences, we just return them all for now. This happens _very_ infrequently.

A better approach would be to apply some kind of selection without dimensionality reduction, but this should work for now.